flow:pwm: Fix possible overflow
Problem spotted by coverity:

Potentially overflowing expression (val.val - val.min) * period with
type int (32 bits, signed) is evaluated using 32-bit arithmetic, and
then used in a context that expects an expression of type
int64_t (64 bits, signed).

To avoid overflow, cast either val.val - val.min or period to type
int64_t.

+++ src/modules/flow/pwm/pwm.c
@@ -86,7 +86,7 @@ map_irange_to_period(struct sol_irange val, int32_t period)
         return UINT32_MAX;
     }
 
-    result = (val.val - val.min) * period / (val.max - val.min);
+    result = (val.val - val.min) * (int64_t)period / (val.max - val.min);
     SOL_INT_CHECK(result, < 0, 0);
 
     return (uint32_t)result;
